+++ client/src/store/modules/tasks.js
@@ -28,7 +28,7 @@ const getters = {
 
       const isForToday = !isAbandoned && plannedAtDate.isBetween(today.startOf('day'), today.endOf('day'), 'day', '[]')
       const isBacklogged = !isFinished && !isAbandoned && !isForToday
-      const createdSinceWeeks = task.createdAt > 0 ? moment.utc().diff(moment.unix(task.createdAt), 'weeks') : 0
+      const startedSinceWeeks = task.startedAt > 0 ? moment.utc().diff(moment.unix(task.startedAt), 'weeks') : 0
 
       const allowedTags = ['b', 'i', 'em', 'strong']
       const anchorOptions = {
@@ -45,7 +45,7 @@ const getters = {
         isBacklogged,
         isFinished,
         isAbandoned,
-        createdSinceWeeks,
+        startedSinceWeeks,
         restartedCount: task.plannedCount - 1,
         plannedAtLabel: formatDate(task.plannedAt),
         formattedLabel: anchorme(sanitizeHtml(task.label, { allowedTags }), anchorOptions),
@@ -104,8 +104,8 @@ const getters = {
   countCreatedByDays (state, getters) {
     const byDays = {}
     getters.list.forEach((task) => {
-      const createdAt = moment.unix(task.createdAt)
-      const key = createdAt.format('YYYY-MM-DD')
+      const startedAt = moment.unix(task.startedAt)
+      const key = startedAt.format('YYYY-MM-DD')
       byDays[key] = (byDays[key] || 0) + 1
     })
     return byDays